+++ src/it/scala/co/topl/brambl/cli/BaseConstants.scala
@@ -13,6 +13,8 @@ trait BaseConstants {
 
   val BASE_AMOUNT = 1000
 
+  val BASE_FEE = 10
+
   val HOST = "localhost"
 
   val BIFROST_PORT = 9084